# api_gateway/api/docs/swagger.yaml
definitions:
  http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
...
  models.SupportTeacherPanelResponse:
    properties:
      groups:
        items:
          $ref: '#/definitions/schedule_service.Group'
        type: array
    type: object
...
  schedule_service.Group:
    properties:
      branch_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      jurnal:
        $ref: '#/definitions/schedule_service.Jurnal'
      students:
        items:
          $ref: '#/definitions/schedule_service.StudentForRep'
        type: array
      support_teacher_id:
        type: string
      teacher_id:
        type: string
      type:
        type: string
      uniqueID:
        type: string
      updated_at:
        type: string
    type: object
  schedule_service.Jurnal:
    properties:
      Schedules:
        items:
          $ref: '#/definitions/schedule_service.Schedule'
        type: array
      Students:
        items:
          $ref: '#/definitions/schedule_service.StudentForRep'
        type: array
      created_at:
        type: string
      from:
        type: string
      group_id:
        type: string
      id:
        type: string
      to:
        type: string
      updated_at:
        type: string
    type: object
  schedule_service.Lesson:
    properties:
      Tasks:
        items:
          $ref: '#/definitions/schedule_service.Task'
        type: array
      created_at:
        type: string
      id:
        type: string
      lesson:
        type: string
      schedule_id:
        type: string
      updated_at:
        type: string
    type: object
  schedule_service.Schedule:
    properties:
      Lesson:
        $ref: '#/definitions/schedule_service.Lesson'
      created_at:
        type: string
      date:
        type: string
      end_time:
        type: string
      id:
        type: string
      journal_id:
        type: string
      start_time:
        type: string
      updated_at:
        type: string
    type: object
...
  schedule_service.StudentForRep:
    properties:
      branch_id:
        type: string
      created_at:
        type: string
      full_name:
        type: string
      group_id:
        type: string
      id:
        type: string
      login:
        type: string
      password:
        type: string
      phone:
        type: string
      role_id:
        type: string
      updated_at:
        type: string
    type: object
  schedule_service.Task:
    properties:
      created_at:
        type: string
      deadline:
        type: string
      id:
        type: string
      label:
        type: string
      lesson_id:
        type: string
      score:
        type: integer
      updated_at:
        type: string
    type: object
...
  user_service.CreateTeacher:
    properties:
      branch_id:
        type: string
      full_name:
        type: string
      ielts_score:
        type: string
      password:
        type: string
      phone:
        type: string
      salary:
        type: number
    type: object
...
  user_service.GetListTeacherResponse:
    properties:
      Teachers:
        items:
          $ref: '#/definitions/user_service.Teacher'
        type: array
      count:
        type: integer
    type: object
...
  user_service.SupportTeacher:
    properties:
      branch_id:
        type: string
      created_at:
        type: string
      full_name:
        type: string
      id:
        type: string
      ielts_score:
        type: string
      login:
        type: string
      password:
        type: string
      phone:
        type: string
      role_id:
        type: string
      salary:
        type: number
      updated_at:
        type: string
    type: object
  user_service.Teacher:
    properties:
      branch_id:
        type: string
      created_at:
        type: string
      full_name:
        type: string
      id:
        type: string
      ielts_score:
        type: string
      login:
        type: string
      password:
        type: string
      phone:
        type: string
      role_id:
        type: string
      salary:
        type: number
      updated_at:
        type: string
    type: object
...
  user_service.UpdateTeacher:
    properties:
      branch_id:
        type: string
      full_name:
        type: string
      id:
        type: string
      ielts_score:
        type: string
      login:
        type: string
      password:
        type: string
      phone:
        type: string
      salary:
        type: number
    type: object
info:
  contact: {}
paths:
...
  /support-teacher:
    get:
      consumes:
      - application/json
      description: Get Teachers List
      operationId: get_Teacher_list
      parameters:
      - description: role_id
        in: query
        name: role_id
        required: true
        type: string
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      - default: a1924766-a9ee-11ed-afa1-0242ac120001
        description: Platform-Id
        in: header
        name: Platform-Id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetAllTeacherResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/user_service.GetListTeacherResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get Teachers List
      tags:
      - SupportTeacher
    post:
      consumes:
      - application/json
      description: Create SupportTeacher
      operationId: create_SupportTeacher
      parameters:
      - description: role_id
        in: query
        name: role_id
        required: true
        type: string
      - description: CreateTeacherBody
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/user_service.CreateTeacher'
      produces:
      - application/json
      responses:
        "200":
          description: GetTeacherBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/user_service.SupportTeacher'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create SupportTeacher
      tags:
      - SupportTeacher
  /support-teacher-panel/{support_teacher_id}:
    get:
      consumes:
      - application/json
      description: Panel Teacher
      operationId: support_teacher_panel
      parameters:
      - description: support_teacher_id
        in: query
        name: support_teacher_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: SupportTeacherPanelResponse
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SupportTeacherPanelResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: support_teacher_panel
      tags:
      - SupportTeacher
  /support-teacher/{id}:
    delete:
      consumes:
      - application/json
      description: Delete SupportTeacher
      operationId: delete_teacher
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: role_id
        in: query
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: SupportTeacher data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete SupportTeacher
      tags:
      - SupportTeacher
    get:
      consumes:
      - application/json
      description: Get SupportTeacher  By ID
      operationId: get_SupportTeacher_by_id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: role_id
        in: query
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: TeacherBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/user_service.SupportTeacher'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get SupportTeacher  By ID
      tags:
      - SupportTeacher
    put:
      consumes:
      - application/json
      description: Update SupportTeacher
      operationId: update_Teacher
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: role_id
        in: query
        name: role_id
        required: true
        type: string
      - description: UpdateTeacher
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/user_service.UpdateTeacher'
      produces:
      - application/json
      responses:
        "200":
          description: SupportTeacher data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/user_service.SupportTeacher'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update SupportTeacher
      tags:
      - SupportTeacher
...
  /teacher:
    get:
      consumes:
      - application/json
      description: Get Teachers List
      operationId: get_teacher_list
      parameters:
      - description: role_id
        in: query
        name: role_id
        required: true
        type: string
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      - default: a1924766-a9ee-11ed-afa1-0242ac120001
        description: Platform-Id
        in: header
        name: Platform-Id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetAllTeacherResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/user_service.GetListTeacherResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get Teachers List
      tags:
      - Teacher
    post:
      consumes:
      - application/json
      description: Create Teacher
      operationId: create_Teacher
      parameters:
      - description: role_id
        in: query
        name: role_id
        required: true
        type: string
      - description: CreateTeacherBody
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/user_service.CreateTeacher'
      produces:
      - application/json
      responses:
        "200":
          description: GetTeacherBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/user_service.Teacher'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Teacher
      tags:
      - Teacher
  /teacher/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Teacher
      operationId: delete_Teacher
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: role_id
        in: query
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Teacher data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Teacher
      tags:
      - Teacher
    get:
      consumes:
      - application/json
      description: Get Teacher  By ID
      operationId: get_Teacher_by_id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: role_id
        in: query
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: TeacherBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/user_service.Teacher'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Teacher  By ID
      tags:
      - Teacher
    put:
      consumes:
      - application/json
      description: Update Teacher
      operationId: update_teacher
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: role_id
        in: query
        name: role_id
        required: true
        type: string
      - description: UpdateTeacher
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/user_service.UpdateTeacher'
      produces:
      - application/json
      responses:
        "200":
          description: Teacher data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/user_service.Teacher'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Teacher
      tags:
      - Teacher
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"